notifier_spec: get rid of duplicate tests
Also:

* factor out blacklist/whitelist tests to separate files (it's
  easier to maintain them that way).
* refactor some tests to avoid deep nesting
* standardise some naming of variables inside tests
